+++ ft_putuns.c
@@ -12,9 +12,9 @@
 
 #include "ft_printf.h"
 
-static int 	uns_count (unsigned int num)
+static ssize_t	uns_count(unsigned int num)
 {
-	int				size;
+	size_t			size;
 	unsigned long	n;
 
 	size = 0;
@@ -41,7 +41,7 @@ static void	ft_sub_uns(unsigned int num)
 		ft_putchar(n + '0');
 }
 
-int ft_putuns (unsigned int num)
+ssize_t	ft_putuns(unsigned int num)
 {
 	int	size;
